Fix password visibility toggle positioning
While testing on Epiphany (GNOME webkit based browser), the password
show toggle appears to be weirdly positioned lower than it should.
This inaccuracy can easily be avoided by not positioning the element
relative to the outer field container, by rather providing a closer
container to closely position it relative to the input field.