Fix password match validation bug
The introduction of a wrapper element for the password visibility toggle
positioning (#7a5ac78) resulted in an additional level of depth in the
DOM subtree responsible for the validation message field.

This resulted in the used querySelector being undefined and throwing an
error when trying to update the relevant element.